changed structure and clean up
