+++ src/main/java/org/matsim/run/RunBerlinScenario.java
@@ -96,9 +96,11 @@ public final class RunBerlinScenario {
 //		DynamicShutdownConfigGroup dynamicShutdownConfigGroup= ConfigUtils.addOrGetModule(config, DynamicShutdownConfigGroup.class);
 
 		DynamicShutdownConfigGroup dynamicShutdownConfigGroup = new DynamicShutdownConfigGroup();
+		dynamicShutdownConfigGroup.setDynamicShutdownModuleActive(DynamicShutdownConfigGroup.dynamicShutdownOptions.ON_FULL);
 		dynamicShutdownConfigGroup.setModeThreshold(1);
 		dynamicShutdownConfigGroup.setIterationToStartFindingSlopes(3);
 		dynamicShutdownConfigGroup.setMinimumWindowSize(3);
+		dynamicShutdownConfigGroup.setMinIterationForGraphics(5);
 		config.addModule(dynamicShutdownConfigGroup);
 
 		//jr end
