+++ src/main/java/org/matsim/run/dynamicShutdown/DynamicShutdownControlerListenerImpl.java
@@ -1,402 +1,502 @@
-//package org.matsim.run.dynamicShutdown;
-//
-//import org.apache.log4j.Logger;
-//import org.matsim.analysis.ModeStatsControlerListener;
-//import org.matsim.analysis.ScoreStats;
-//import org.matsim.analysis.ScoreStatsControlerListener;
-//import org.matsim.api.core.v01.Scenario;
-//import org.matsim.api.core.v01.population.Person;
-//import org.matsim.api.core.v01.population.Plan;
-//import org.matsim.core.config.groups.ControlerConfigGroup;
-//import org.matsim.core.config.groups.StrategyConfigGroup;
-//import org.matsim.core.controler.OutputDirectoryHierarchy;
-//import org.matsim.core.controler.events.IterationStartsEvent;
-//import org.matsim.core.controler.events.ShutdownEvent;
-//import org.matsim.core.controler.events.StartupEvent;
-//import org.matsim.core.controler.listener.IterationStartsListener;
-//import org.matsim.core.controler.listener.ShutdownListener;
-//import org.matsim.core.controler.listener.StartupListener;
-//import org.matsim.core.replanning.*;
-//import org.matsim.core.utils.io.IOUtils;
-//
-//import javax.inject.Inject;
-//import java.io.BufferedWriter;
-//import java.io.IOException;
-//import java.util.*;
-//
-///**
-// * When user-specified criteria are met, initiates dynamic shutdown: 1) Turns off innovation in next iteration and
-// * 2) informs TerminateDynamically Module of the iteration at which MATSim should shut down. \
-// *
-// * The criteria to used must be user specified (i.e. score convergence), as well as the smoothing method and associated
-// * parameters.
-// *
-// * @author jakobrehmann
-// */
-//
-//@Deprecated
-//public class DynamicShutdownControlerListenerImpl implements IterationStartsListener, StartupListener, ShutdownListener {
-//
-//    // Dynamic Shutdown Config Group
-//    private static final int MINIMUM_ITERATION = 10000;
-//
-//    // Score Config Group
-//    private static final boolean SCORE_CONDITION_ACTIVE = true;
-//    private static final double SCORE_CONDITION_THRESHOLD = 0.5;
-//    private static final int SCORE_CONDITION_SMOOTHING_INTERVAL = 50 ;
-//
-//    // Mode Convergence Config Group
-//    private static final boolean MODE_CONDITION_ACTIVE = true;
-//    private static final double MODE_CONDITION_THRESHOLD = 0.0005;
-//    private static final int MODE_CONDITION_SMOOTHING_INTERVAL = 50 ;
-//
-//    // Mode Choice Coverage Config Group
-//    private static final boolean MODECHOICECOVERAGE_CONDITION_ACTIVE = true;
-//    private static final double MODECHOICECOVERAGE_CONDITION_THRESHOLD = 0.0005;
-//    private static final int MODECHOICECOVERAGE_CONDITION_SMOOTHING_INTERVAL = 50 ;
-//
-//
-//    private int lastIteration;
-//    private double fractionOfIterationsToDisableInnovation;
-//    private final ControlerConfigGroup controlerConfigGroup;
-//    private final OutputDirectoryHierarchy controlerIO;
-//    private Scenario scenario;
-//    private ScoreStats scoreStats;
-//    private ModeStatsControlerListener modeStats;
-//    ModeStatsControlerListener modeStatsControlerListener;
-//
-//    private StrategyManager strategyManager;
-//    private static int dynamicShutdownIteration;
-//    private static boolean dynamicShutdownInitiated;
-//    private StrategyConfigGroup strategyConfigGroup;
-//    private static final Logger log = Logger.getLogger(StrategyManager.class);
-//
-//
-//    private static Map<String, List<Double>> pctChangesForScore = new HashMap<>();
-//    private static Map<String, List<Double>> pctChangesForMode = new HashMap<>();
-//    private static Map<String, List<Double>> pctChangesForModeChoiceCoverage = new HashMap<>();
-//
-//
-//
-//    @Inject
-//    DynamicShutdownControlerListenerImpl(ControlerConfigGroup controlerConfigGroup, ScoreStats scoreStats, ModeStatsControlerListener modeStatsControlerListener, StrategyManager strategyManager,
-//                                         StrategyConfigGroup strategyConfigGroup, Scenario scenario, OutputDirectoryHierarchy controlerIO) {
-//
-//        this.scenario = scenario;
-//        this.scoreStats = scoreStats;
-//        this.strategyManager = strategyManager;
-//        this.strategyConfigGroup = strategyConfigGroup;
-//        this.controlerConfigGroup = controlerConfigGroup ;
-//        this.controlerIO = controlerIO ;
-//        this.modeStatsControlerListener = modeStatsControlerListener ;
-//    }
-//
-//    public static int getDynamicShutdownIteration() {
-//        return dynamicShutdownIteration;
-//    }
-//
-//    public static boolean isDynamicShutdownInitiated() {
-//        return dynamicShutdownInitiated;
-//    }
-//
-//    @Override
-//    public void notifyStartup(StartupEvent startupEvent) {
-//        dynamicShutdownInitiated = false ;
-//        dynamicShutdownIteration = controlerConfigGroup.getLastIteration();
-//
-//        fractionOfIterationsToDisableInnovation = strategyConfigGroup.getFractionOfIterationsToDisableInnovation();
-//        lastIteration = scenario.getConfig().controler().getLastIteration();
-//    }
-//
-//    @Override
-//    public void notifyIterationStarts(IterationStartsEvent iterationStartsEvent) {
-//        int iteration = iterationStartsEvent.getIteration();
-//
-//        // Checks 1:  at least one shutdown criteria is being active. Otherwise this module is superfluous.
-//        if (!MODECHOICECOVERAGE_CONDITION_ACTIVE && !SCORE_CONDITION_ACTIVE && !MODE_CONDITION_ACTIVE) {
-//            log.warn("dynamic shutdown should not be used if no criteria are specified. " +
-//                    "Therefore, the standard shutdown iteration will be used");
-//            return;
-//        }
-//
-//
-//        // Check 2: checks whether dynamic shutdown was already initiated.
-//        if (dynamicShutdownInitiated) {
-//            log.info("dynamic shutdown was previously initiated");
-//            return;
-//        }
-//
-//        // jr temporary:
-//        pctChangeForModeChoiceCoverage(iteration);
-//        pctChangeForScore(iteration);
-////        pctChangeForMode(iteration);
-////
-//        // Check 3: checks whether the minimum iteration was reached ;
-//        if (iteration < MINIMUM_ITERATION) {
-//            return ;
-//        }
-//
-//
-////        // Check 4: check whether innovation shutdown has already occured
-////        if (iteration >= fractionOfIterationsToDisableInnovation * lastIteration) {
-////            return;
-////        }
-//
-//        // Check 5: returns if the mode choice coverage criteria has not yet been met (assuming the criteria is active)
-//        //     step A: Add newest percent difference to the pctDifference ArrayList
-//        //     step B: Check Last x percent changes, to see if innovation shutdown should be initiated
-//        if (MODECHOICECOVERAGE_CONDITION_ACTIVE) {
-////            pctChangeForModeChoiceCoverage(iteration);
-//            if (!checkModeChoiceCoverageCriteria(MODECHOICECOVERAGE_CONDITION_SMOOTHING_INTERVAL)) {
-//                return;
-//            }
-//        }
-//
-//        // Check 6: returns if the score criteria has not yet been met (assuming the criteria is active)
-//        if (SCORE_CONDITION_ACTIVE) {
-////            pctChangeForScore(iteration);
-//            if (!checkScoreCriteria(SCORE_CONDITION_SMOOTHING_INTERVAL)) {
-//                return;
-//            }
-//        }
-//
-//        // Mode Convergence - work in progress
-//        if (MODE_CONDITION_ACTIVE) {
-////            pctChangeForMode(iteration);
-////            if (!checkModeCriteria(MODE_CONDITION_SMOOTHING_INTERVAL)) {
-////                return;
-////            }
-//        }
-//
-//        // FINALLY: if none of the previous checks terminated the process, then dynamic shutdown can be initiated.
-//        shutdownInnovation(iteration);
-//    }
-//
-//
-////    private void pctChangeForMode(int iteration) {
-////        Map<String, Map<Integer, Double>> modeHistory;
-////        try {
-////
-//////            modeHistory = modeStats.getModeHistories();
-////
-////
-////            for (Map.Entry<String, Map<Integer, Double>> entry : modeHistory.entrySet()) {
-////                String mode = entry.getKey();
-////                log.info("Mode checked for " + mode);
-////                Map<Integer, Double> map = entry.getValue();
-////
-////                Double val1 = map.get(iteration - 2);
-////                System.out.println(val1);
-////
-////                Double val2 = map.get(iteration - 1);
-////                System.out.println(val2);
-////
-////                Double pctChange = Math.abs((val1 - val2) / val2);
-////                List<Double> list = pctChangesForMode.computeIfAbsent(mode, v -> new ArrayList<>());
-////                list.add(pctChange);
-////            }
-////        } catch (NullPointerException e) {
-////            log.error("Too early to find percent change");
-////        }
-////
-////    }
-//
-//
-//    private void pctChangeForModeChoiceCoverage(int iteration) {
-//        Map<Integer, Map<String, Map<Integer, Double>>> modeChoiceCoverageHistory;
-//        try {
-//
-//            modeChoiceCoverageHistory = ModeChoiceCoverageControlerListener.getModeChoiceCoverageHistory();
-//
-//            Integer limit = 1;
-//
-//            for (Map.Entry<String, Map<Integer, Double>> entry : modeChoiceCoverageHistory.get(limit).entrySet()) {
-//                String mode = entry.getKey();
-//                log.info("Mode choice coverage checked for " + mode);
-//                Map<Integer, Double> map = entry.getValue();
-//
-//                Double val1 = map.get(iteration - 2);
-//                System.out.println(val1);
-//
-//                Double val2 = map.get(iteration - 1);
-//                System.out.println(val2);
-//
-//                Double pctChange = Math.abs((val1 - val2) / val2);
-//                List<Double> pctChangeForMode = pctChangesForModeChoiceCoverage.computeIfAbsent(mode, v -> new ArrayList<>());
-//                pctChangeForMode.add(pctChange);
-//            }
-//        } catch (NullPointerException e) {
-//            log.error("Too early to find percent change");
-//        }
-//
-//    }
-//
-//
-//    private boolean checkModeChoiceCoverageCriteria(int dx) {
-//
-//        // Calculates Average Percent Change for all modes
-//        Map<String, Double> averageModeChoiceCoveragePctChanges = new HashMap<>();
-//        for (String mode : pctChangesForModeChoiceCoverage.keySet()) {
-//
-//
-//            List<Double> pctChangesForMode = pctChangesForModeChoiceCoverage.get(mode);
-//            if (dx > pctChangesForMode.size()) {
-//                log.warn("An average percent change over cannot yet be found, since dx is too high.");
-//                return false;
-//            }
-//            double avgPctChng = averagePercentChange(pctChangesForMode, dx);
-//            averageModeChoiceCoveragePctChanges.put(mode, avgPctChng);
-//            System.out.println("average percent change (dx=1) for mode " + mode + " = " + avgPctChng + ", satisfies precision threshhold: " + (avgPctChng <= MODECHOICECOVERAGE_CONDITION_THRESHOLD));
-//
-//        }
-//
-//
-//        // Returns false if at least one mode is above threshhold
-//        for (Double pc : averageModeChoiceCoveragePctChanges.values()) {
-//            if (pc > MODECHOICECOVERAGE_CONDITION_THRESHOLD) {
-//                return false;
-//            }
-//        }
-//
-//        return true;
-//    }
-//
-//    private void pctChangeForScore(int iteration) {
-//        try {
-//            for (Map.Entry<ScoreStatsControlerListener.ScoreItem, Map<Integer, Double>> entry : scoreStats.getScoreHistory().entrySet()) {
-//                String scoreItem = entry.getKey().toString();
-//                log.info("Score Percent Change  checked for " + scoreItem);
-//                Map<Integer, Double> map = entry.getValue();
-//
-//                Double val1 = map.get(iteration - 2);
-//                System.out.println(val1);
-//
-//                Double val2 = map.get(iteration - 1);
-//                System.out.println(val2);
-//
-//                Double pctChange = Math.abs((val1 - val2) / val2);
-//                List<Double> pctChangeForScoreItem = pctChangesForScore.computeIfAbsent(scoreItem, v -> new ArrayList<>());
-//                pctChangeForScoreItem.add(pctChange);
-//            }
-//        } catch (NullPointerException e) {
-//            log.error("Too early to find percent change (score stats)");
-//        }
-//    }
-//
-//    private boolean checkScoreCriteria(int dx) {
-//
-//        // Calculates Average Percent Change for all score items
-//        Map<String, Double> averageScorePctChanges = new HashMap<>();
-//        for (String scoreItem : pctChangesForScore.keySet()) {
-//
-//            if (dx > pctChangesForScore.get(scoreItem).size()){
-//                log.warn("An average percent change over cannot yet be found, since dx is too high.");
-//                return false;
-//            }
-//
-//            double avgPctChng = averagePercentChange(pctChangesForScore.get(scoreItem), dx);
-//            averageScorePctChanges.put(scoreItem, avgPctChng);
-//            System.out.println("average percent change (dx= "+ SCORE_CONDITION_SMOOTHING_INTERVAL
-//                    + ") for score item " + scoreItem + " = " + avgPctChng + ", satisfies precision threshhold: "
-//                    + (avgPctChng <= SCORE_CONDITION_THRESHOLD));
-//
-//        }
-//
-//
-//        // Returns false if at least one mode is above threshhold
-//        for (Double pc : averageScorePctChanges.values()) {
-//            if (pc > SCORE_CONDITION_THRESHOLD) {
-//                return false;
-//            }
-//        }
-//
-//        return true;
-//    }
-//
-//    private double averagePercentChange(List<Double> list, int dx) {
-//
-//        double avgPctChange = 0.;
-//
-//        for (int i = list.size() - 1; i >= list.size() - dx; i--) {
-//            avgPctChange += list.get(i);
-//        }
-//
-//        return avgPctChange / dx;
-//
-//    }
-//
-//    private void shutdownInnovation(int iteration) {
-//        dynamicShutdownIteration = (int) (iteration / fractionOfIterationsToDisableInnovation) + 2; // jr review
-//        int nextDisableInnovativeStrategiesIteration = iteration + 1; // jr review
-//
-//
-//        Set<String> subpopulations = new HashSet<>();
-//        for (StrategyConfigGroup.StrategySettings setting : this.scenario.getConfig().strategy().getStrategySettings()) {
-//            subpopulations.add(setting.getSubpopulation());
-//            if (subpopulations.size() == 0) subpopulations.add(null);
-//        }
-//
-//        for (String subpopulation : subpopulations) {
-//            for (GenericPlanStrategy<Plan, Person> planStrategy : strategyManager.getStrategies(subpopulation)) {
-//                PlanStrategyImpl planStrategyImpl = (PlanStrategyImpl) planStrategy;
-//                if (isInnovativeStrategy(planStrategyImpl)) {
-////                        log.info("Setting weight for " + planStrategyImpl.toString() + " (subpopulation " + subpopulation + ") to 0.");
-//                    strategyManager.addChangeRequest(nextDisableInnovativeStrategiesIteration, planStrategyImpl, subpopulation, 0.);
-//
-//
-//                }
-//            }
-//        }
-//
-//        log.error("Innovation Shutdown at iteration " + (nextDisableInnovativeStrategiesIteration));
-//        log.error("Full Shutdown at iteration " + dynamicShutdownIteration);
-//        dynamicShutdownInitiated = true;
-//    }
-//
-//
-//    private boolean isInnovativeStrategy(GenericPlanStrategy<Plan, Person> strategy) {
-//        return !(ReplanningUtils.isOnlySelector(strategy));
-//    }
-//
-//    @Override
-//    public void notifyShutdown(ShutdownEvent shutdownEvent) {
-//
-//
-//        // mode choice coverage
-//        try (BufferedWriter bw = IOUtils.getBufferedWriter(controlerIO.getOutputFilename("PercentDifferencesModeChoiceCoverage.txt"))){
-//            for (String mode : pctChangesForModeChoiceCoverage.keySet()) {
-//                bw.write("\n" + mode+" ; ");
-//                for (Double pct : pctChangesForModeChoiceCoverage.get(mode)) {
-//                    bw.write(pct + " ; ");
-//                }
-//            }
-//        } catch (IOException e) {
-//            e.printStackTrace();
-//        }
-//
-//        // score
-//        try (BufferedWriter bw = IOUtils.getBufferedWriter(controlerIO.getOutputFilename("PercentDifferencesScore.txt"))){
-//            for (String scoreStat : pctChangesForScore.keySet()) {
-//                bw.write("\n"+scoreStat+" ; ");
-//                for (Double pct : pctChangesForScore.get(scoreStat)) {
-//                    bw.write(pct + " ; ");
-//                }
-//            }
-//        } catch (IOException e) {
-//            e.printStackTrace();
-//        }
-//
-//        // mode
-////        try (BufferedWriter bw = IOUtils.getBufferedWriter(controlerIO.getOutputFilename("PercentDifferencesMode.txt"))){
-////            for (String mode : pctChangesForMode.keySet()) {
-////                bw.write("\n"+mode+" ; ");
-////                for (Double pct : pctChangesForMode.get(mode)) {
-////                    bw.write(pct + " ; ");
-////                }
-////            }
-////        } catch (IOException e) {
-////            e.printStackTrace();
-////        }
-//    }
-//}
-//
+package org.matsim.run.dynamicShutdown;
+
+import org.apache.log4j.Logger;
+import org.matsim.analysis.ModeStatsControlerListener;
+import org.matsim.analysis.ScoreStats;
+import org.matsim.analysis.ScoreStatsControlerListener.ScoreItem;
+import org.matsim.api.core.v01.Scenario;
+import org.matsim.api.core.v01.population.Person;
+import org.matsim.api.core.v01.population.Plan;
+import org.matsim.core.config.groups.ControlerConfigGroup;
+import org.matsim.core.config.groups.PlanCalcScoreConfigGroup;
+import org.matsim.core.config.groups.StrategyConfigGroup;
+import org.matsim.core.controler.OutputDirectoryHierarchy;
+import org.matsim.core.controler.events.IterationStartsEvent;
+import org.matsim.core.controler.events.ShutdownEvent;
+import org.matsim.core.controler.events.StartupEvent;
+import org.matsim.core.controler.listener.IterationStartsListener;
+import org.matsim.core.controler.listener.ShutdownListener;
+import org.matsim.core.controler.listener.StartupListener;
+import org.matsim.core.replanning.GenericPlanStrategy;
+import org.matsim.core.replanning.PlanStrategyImpl;
+import org.matsim.core.replanning.ReplanningUtils;
+import org.matsim.core.replanning.StrategyManager;
+import org.matsim.core.utils.io.IOUtils;
+import org.matsim.core.utils.io.UncheckedIOException;
+
+import javax.inject.Inject;
+import java.io.BufferedWriter;
+import java.io.IOException;
+import java.util.*;
+
+import static org.matsim.run.dynamicShutdown.DynamicShutdownConfigGroup.slopeWindowOption.EXPANDING;
+
+/**
+ * When user-specified criteria are met, dynamic shutdown is initiated: 1) Turns off innovation in next iteration and
+ * 2) informs TerminateDynamically Module of the iteration at which MATSim should shut down.
+ *
+ * The criteria to used must be user specified (i.e. score convergence), as well as the smoothing method and associated
+ * parameters.
+ *
+ * @author jakobrehmann
+ */
+
+
+public class DynamicShutdownControlerListenerImpl implements IterationStartsListener, StartupListener, ShutdownListener, DynamicShutdownControlerListener {
+
+    private final BufferedWriter slopesOut ;
+    private final ControlerConfigGroup controlerConfigGroup;
+    private final Scenario scenario;
+    private final ScoreStats scoreStats;
+    private final ModeStatsControlerListener modeStatsControlerListener;
+    private final ModeChoiceCoverageControlerListener modeChoiceCoverageControlerListener;
+    private String outputFileName;
+    private final int globalInnovationDisableAfter;
+    private final PlanCalcScoreConfigGroup scoreConfig;
+
+
+    private final StrategyManager strategyManager;
+    private static int dynamicShutdownIteration;
+    private static boolean dynamicShutdownInitiated;
+    private final StrategyConfigGroup strategyConfigGroup;
+    private static final Logger log = Logger.getLogger(StrategyManager.class);
+
+
+    private static final Map<String, Map<Integer,Double>> slopesScore = new HashMap<>();
+    private static final Map<String, Map<Integer,Double>> slopesMode = new HashMap<>();
+    private static final Map<String, Map<Integer,Double>> slopesModeChoiceCoverage = new HashMap<>();
+
+
+    private static final Map<String, Map<Integer,Integer>> convergenceScore = new HashMap<>();
+    private static final Map<String, Map<Integer,Integer>> convergenceMode = new HashMap<>();
+    private static final Map<String, Map<Integer,Integer>> convergenceModeCC = new HashMap<>();
+
+
+    private List<String> activeMetricsScore = new ArrayList<>();
+    private List<String> activeMetricsMode = new ArrayList<>();
+    private List<String> activeMetricsModeCC = new ArrayList<>();
+
+    private DynamicShutdownConfigGroup cfg;
+    private int innoShutoffIter;
+
+
+    @Inject
+    DynamicShutdownControlerListenerImpl(ControlerConfigGroup controlerConfigGroup, ScoreStats scoreStats,
+                                         ModeStatsControlerListener modeStatsControlerListener, StrategyManager strategyManager,
+                                         StrategyConfigGroup strategyConfigGroup, Scenario scenario, OutputDirectoryHierarchy controlerIO,
+                                        PlanCalcScoreConfigGroup scoreConfig,
+                                         ModeChoiceCoverageControlerListener modeChoiceCoverageControlerListener) {
+
+        this.scenario = scenario;
+        this.scoreStats = scoreStats;
+        this.strategyManager = strategyManager;
+        this.strategyConfigGroup = strategyConfigGroup;
+        this.controlerConfigGroup = controlerConfigGroup ;
+        this.modeStatsControlerListener = modeStatsControlerListener ;
+        this.outputFileName = controlerIO.getOutputFilename("dynShutdown_");
+        this.modeChoiceCoverageControlerListener = modeChoiceCoverageControlerListener;
+
+        this.cfg = (DynamicShutdownConfigGroup) scenario.getConfig().getModules().get(DynamicShutdownConfigGroup.GROUP_NAME);
+
+        this.globalInnovationDisableAfter = (int) ((controlerConfigGroup.getLastIteration() - controlerConfigGroup.getFirstIteration())
+                * strategyConfigGroup.getFractionOfIterationsToDisableInnovation() + controlerConfigGroup.getFirstIteration());
+
+        this.slopesOut = IOUtils.getBufferedWriter(this.outputFileName + "allMetrics.txt");
+
+        this.scoreConfig = scoreConfig;
+    }
+
+
+
+    @Override
+    public int getDynamicShutdownIteration() {
+        return dynamicShutdownIteration;
+    }
+
+    @Override
+    public boolean isDynamicShutdownInitiated() {
+        return dynamicShutdownInitiated;
+    }
+
+    @Override
+    public void notifyStartup(StartupEvent startupEvent) {
+        dynamicShutdownInitiated = false ;
+        dynamicShutdownIteration = Integer.MAX_VALUE;
+        innoShutoffIter = Integer.MAX_VALUE;
+
+        generateMetricLists(scoreConfig);
+
+        try {
+            this.slopesOut.write("Iteration");
+
+            for ( String scoreType : activeMetricsScore) {
+                this.slopesOut.write("\tscore-" + scoreType+ "\tconverged");
+            }
+            for ( String mode : activeMetricsMode) {
+                this.slopesOut.write("\tmode-" + mode+ "\tconverged");
+            }
+            for ( String mode : activeMetricsModeCC) {
+                this.slopesOut.write("\tmodeCC-" + mode + "\tconverged");
+            }
+            this.slopesOut.write("\tnotes\n");
+            this.slopesOut.flush();
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+
+    }
+
+    @Override
+    public void notifyIterationStarts(IterationStartsEvent iterationStartsEvent) {
+
+        int iteration = iterationStartsEvent.getIteration();
+        int prevIteration = iteration - 1;
+
+        if (cfg.isDynamicShutdownModuleActive()== DynamicShutdownConfigGroup.dynamicShutdownOptions.OFF) {
+            log.warn("Dynamic shutdown module not active");
+        }
+        
+        if (activeMetricsScore.isEmpty() && activeMetricsMode.isEmpty() && activeMetricsModeCC.isEmpty()) {
+            log.warn("No criteria specified for dynamic shutdown module");
+            return;
+        }
+        
+        if (iteration < cfg.getIterationToStartFindingSlopes()) {
+            return;
+        }
+
+        try {
+            this.slopesOut.write("\n" + prevIteration);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+        boolean scoreConverged = false;
+        if (!activeMetricsScore.isEmpty()) {
+            String metricType = "score";
+            Map<ScoreItem, Map<Integer, Double>> scoreHistory = scoreStats.getScoreHistory();
+            Map<String, Map<Integer, Double>> scoreHistoryMod = new HashMap<>();
+            for (ScoreItem scoreItem : scoreHistory.keySet()) {
+                scoreHistoryMod.put(scoreItem.name(), scoreHistory.get(scoreItem));
+            }
+
+            bestFitLineGeneric(iteration, scoreHistoryMod, slopesScore, activeMetricsScore, metricType);
+            produceDynShutdownGraphs(scoreHistoryMod, slopesScore, metricType, activeMetricsScore, cfg.getScoreThreshold(), iteration);
+
+            scoreConverged = metricTypeConverges(slopesScore, convergenceScore, activeMetricsScore, cfg.getScoreThreshold(), prevIteration);
+
+            writeSlopeAndConvergence(slopesScore, convergenceScore, activeMetricsScore, prevIteration);
+
+        }
+
+        boolean modeConverged = false;
+        if (!activeMetricsMode.isEmpty()) {
+            String metricType = "mode";
+            Map<String, Map<Integer, Double>> modeHistories = modeStatsControlerListener.getModeHistories();
+            bestFitLineGeneric(iteration, modeHistories, slopesMode, activeMetricsMode, metricType);
+            produceDynShutdownGraphs(modeHistories, slopesMode, metricType, activeMetricsMode, cfg.getModeThreshold(), iteration);
+
+            modeConverged = metricTypeConverges(slopesMode, convergenceMode, activeMetricsMode, cfg.getModeThreshold(), prevIteration);
+
+            writeSlopeAndConvergence(slopesMode, convergenceMode, activeMetricsMode, prevIteration);
+        }
+
+        boolean modeCCConverged = false;
+        if (!activeMetricsModeCC.isEmpty()) {
+
+            String metricType = "modeChoiceCoverage";
+            int mCCLimit = 1;
+            Map<String, Map<Integer, Double>> mCCHistory = modeChoiceCoverageControlerListener.getModeChoiceCoverageHistory().get(mCCLimit);
+            bestFitLineGeneric(iteration, mCCHistory, slopesModeChoiceCoverage, activeMetricsModeCC, metricType);
+            produceDynShutdownGraphs(mCCHistory,slopesModeChoiceCoverage, metricType, activeMetricsModeCC, cfg.getModechoicecoverageThreshold(),iteration);
+
+            modeCCConverged = metricTypeConverges(slopesModeChoiceCoverage, convergenceModeCC, activeMetricsModeCC, cfg.getModechoicecoverageThreshold(), prevIteration);
+
+            writeSlopeAndConvergence(slopesModeChoiceCoverage, convergenceModeCC, activeMetricsModeCC, prevIteration);
+        }
+
+
+        try {
+            if (iteration == innoShutoffIter) {
+                this.slopesOut.write("\tNOTE: innovation turned off");
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+
+        if (cfg.isDynamicShutdownModuleActive() == DynamicShutdownConfigGroup.dynamicShutdownOptions.ON_ANALYSIS_ONLY) {
+            return;
+        }
+
+        if (dynamicShutdownInitiated) {
+            log.info("Dynamic shutdown was previously initiated");
+            return;
+        }
+
+        if (iteration < cfg.getMinimumIteration()) {
+            return ;
+        }
+
+        if (iteration >= globalInnovationDisableAfter) {
+            return;
+        }
+
+        if (!activeMetricsScore.isEmpty() && !scoreConverged) {
+            return;
+        }
+
+        if (!activeMetricsMode.isEmpty() && !modeConverged) {
+            return;
+        }
+
+        if (!activeMetricsModeCC.isEmpty() && !modeCCConverged) {
+            return;
+        }
+
+
+        shutdownInnovation(iteration);
+
+    }
+
+    private void writeSlopeAndConvergence(Map<String, Map<Integer, Double>> slopesMap,
+                                          Map<String, Map<Integer, Integer>> convergenceMap,
+                                          List<String> metricsToInclude,int prevIteration) {
+        try{
+            for (String metric : metricsToInclude) {
+                Double slope = slopesMap.get(metric).get(prevIteration);
+
+                int convCnt = convergenceMap.get(metric).get(prevIteration);
+                String convStr;
+                if (convCnt >= cfg.getIterationsInZoneToConverge()) {
+                    convStr = "true";
+                } else {
+                    convStr = convCnt + "/" + cfg.getIterationsInZoneToConverge();
+                }
+                this.slopesOut.write("\t"+ slope + "\t" + convStr);
+            }
+            this.slopesOut.flush();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+
+    private boolean metricTypeConverges(Map<String, Map<Integer, Double>> slopesMap,
+                                       Map<String, Map<Integer, Integer>> convergenceMap,
+                                       List<String> metricsToInclude,
+                                       double threshold,
+                                       int prevIteration) {
+
+
+
+        if (slopesMap.isEmpty()) {
+            return false;
+        }
+        for (String metric : metricsToInclude) {
+            Map<Integer,Integer> convergenceCntPerMetric = convergenceMap.computeIfAbsent(metric,  v -> new HashMap<>());
+            Double slope = slopesMap.get(metric).get(prevIteration);
+            if (slope > -1 * threshold && slope < threshold) {
+                int convergenceCountSoFar = convergenceCntPerMetric.getOrDefault(prevIteration - 1, 0);
+                convergenceCntPerMetric.put(prevIteration, convergenceCountSoFar + 1);
+            } else {
+                convergenceCntPerMetric.put(prevIteration, 0);
+            }
+        }
+
+
+        for (String metric : metricsToInclude) {
+            if (convergenceMap.get(metric).get(prevIteration) < cfg.getIterationsInZoneToConverge()) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    private void bestFitLineGeneric(int iteration,
+                                    Map<String, Map<Integer, Double>> history,
+                                    Map<String, Map<Integer, Double>> slopes,
+                                    List<String> metricsToInclude,
+                                    String metricType) {
+
+        for (Map.Entry<String, Map<Integer, Double>> entry : history.entrySet()) {
+
+            String metricName = entry.getKey();
+
+            if (!metricsToInclude.isEmpty() &&  !metricsToInclude.contains(metricName)) {
+                log.info(metricType + " NOT checked for " + metricName);
+                continue;
+            }
+
+            log.info(metricType + " checked for " + metricName);
+
+            double slope = computeLineSlope(entry.getValue());
+
+            Map<Integer,Double> slopesForMetric = slopes.computeIfAbsent(metricName, v -> new HashMap<>());
+            slopesForMetric.put(iteration-1,slope); // calculation for the previous iteration
+        }
+    }
+
+
+    private double computeLineSlope(Map<Integer,Double> inputMap) {
+
+
+        int currentIter = Collections.max(inputMap.keySet());
+        int startIteration = currentIter - cfg.getMinimumWindowSize() + 1; // fixed window
+        int startIterationExpanding = (int) ((1.0 - cfg.getExpandingWindowPctRetention()) * currentIter); // expanding window
+        if (cfg.getSlopeWindowPolicy() == EXPANDING && startIterationExpanding < startIteration) { // TODO: Test whether enum works in this case
+            startIteration = startIterationExpanding;
+        }
+
+        ArrayList<Integer> x = new ArrayList<>();
+        ArrayList<Double> y = new ArrayList<>();
+
+        for (Integer it : inputMap.keySet()) {
+            if (it >= startIteration) {
+                x.add(it);
+                y.add(inputMap.get(it));
+            }
+        }
+
+        if (x.size() != y.size()) {
+            throw new IllegalArgumentException("array lengths are not equal");
+        }
+        int n = x.size();
+        System.out.println("XXXXXXXXXXXXXXXX Window size: " + n);
+
+        // first pass
+        double sumx = 0.0, sumy = 0.0;
+        for (int i = 0; i < n; i++) {
+            sumx  += x.get(i);
+            sumy  += y.get(i);
+        }
+        double xbar = sumx / n;
+        double ybar = sumy / n;
+
+        // second pass: compute summary statistics
+        double xxbar = 0.0, xybar = 0.0;
+        for (int i = 0; i < n; i++) {
+            xxbar += (x.get(i) - xbar) * (x.get(i) - xbar);
+            xybar += (x.get(i) - xbar) * (y.get(i) - ybar);
+        }
+        return xybar / xxbar;
+    }
+
+
+    private void produceDynShutdownGraphs(Map<String, Map<Integer, Double>> history,
+                                          Map<String, Map<Integer, Double>> slopes,
+                                          String metricType,
+                                          List<String> metricsToInclude,
+                                          double convergenceThreshold,
+                                          int iteration) {
+
+        if (iteration <= cfg.getMinIterationForGraphics()) {
+            return;
+        }
+
+
+        for (String metricName : metricsToInclude) {
+            try {
+                XYLineChartDualYAxis chart = new XYLineChartDualYAxis("Convergence for " + metricType + " : " + metricName, "iteration", metricType + " : " + metricName, "slope of " + metricName);
+
+                chart.addSeries(metricName, history.get(metricName));
+                chart.addSeries2("slope of " + metricName, slopes.get(metricName));
+
+                chart.addVerticalRange(-convergenceThreshold, convergenceThreshold);
+                chart.addMatsimLogo();
+
+                chart.saveAsPng(outputFileName + metricType + "_" + metricName + ".png", 800, 600);
+            } catch (NullPointerException e) {
+                    log.error("Could not produce Dynamic Shutdown Graphs (probably too early)");
+            }
+        }
+    }
+
+
+    private void shutdownInnovation(int iteration) {
+        innoShutoffIter = iteration + 1; // New weights are in effect in following iteration.
+        double innoPct = strategyConfigGroup.getFractionOfIterationsToDisableInnovation();
+        int firstIter = controlerConfigGroup.getFirstIteration();
+        dynamicShutdownIteration = (int) ((innoShutoffIter - firstIter) / innoPct) + firstIter;
+        dynamicShutdownInitiated = true;
+
+
+        Set<String> subpopulations = new HashSet<>();
+        for (StrategyConfigGroup.StrategySettings setting : this.scenario.getConfig().strategy().getStrategySettings()) {
+            subpopulations.add(setting.getSubpopulation());
+            if (subpopulations.size() == 0) subpopulations.add(null);
+        }
+
+        for (String subpopulation : subpopulations) {
+            for (GenericPlanStrategy<Plan, Person> planStrategy : strategyManager.getStrategies(subpopulation)) {
+                PlanStrategyImpl planStrategyImpl = (PlanStrategyImpl) planStrategy;
+                if (!(ReplanningUtils.isOnlySelector(planStrategyImpl))) { // if (innovation strategy)
+                    strategyManager.addChangeRequest(innoShutoffIter, planStrategyImpl, subpopulation, 0.);
+                }
+            }
+        }
+
+        log.info("********** DYNAMIC SHUTDOWN INITIATED ***********");
+        log.info("Innovation strategies deactivated in iteration " + (innoShutoffIter));
+        log.info("Full shutdown will occur in iteration " + dynamicShutdownIteration);
+
+        try {
+            this.slopesOut.write("\tNOTE: dynamic shutdown initiated");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private void generateMetricLists(PlanCalcScoreConfigGroup scoreConfig) {
+        switch (cfg.getScorePolicyChosen()) {
+            case ON_FULL:
+                activeMetricsScore = new ArrayList<>(Arrays.asList(
+                        ScoreItem.executed.name(),
+                        ScoreItem.average.name(),
+                        ScoreItem.best.name(),
+                        ScoreItem.worst.name()));
+                break;
+            case ON_EXECUTED_ONLY:
+                activeMetricsScore.add(ScoreItem.executed.name());
+                break;
+            case OFF:
+                break;
+            default:
+                break;
+        }
+
+        switch (cfg.getModePolicyChosen()) {
+            case ON_FULL:
+                activeMetricsMode.addAll(scoreConfig.getAllModes());
+                break;
+            case OFF:
+                activeMetricsMode.clear();
+                break;
+            default:
+                activeMetricsMode.clear();
+                break;
+        }
+
+        switch (cfg.getModeCCPolicyChosen()) {
+            case ON_FULL:
+                activeMetricsModeCC.addAll(scoreConfig.getAllModes());
+                break;
+            case OFF:
+                activeMetricsModeCC.clear();
+                break;
+            default:
+                activeMetricsModeCC.clear();
+                break;
+        }
+    }
+
+    @Override
+    public void notifyShutdown(ShutdownEvent shutdownEvent) {
+        try {
+            this.slopesOut.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+}
