+++ src/main/java/org/matsim/run/dynamicShutdown/DynamicShutdownModule.java
@@ -23,7 +23,6 @@
 package org.matsim.run.dynamicShutdown;
 
 import com.google.inject.Singleton;
-import org.matsim.analysis.ModeStatsControlerListener;
 import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.controler.AbstractModule;
 import org.matsim.core.controler.TerminationCriterion;
@@ -39,8 +38,8 @@ public class DynamicShutdownModule extends AbstractModule {
         this.bind(ModeChoiceCoverageControlerListener.class).in(Singleton.class); //TODO: Delete?
         this.addControlerListenerBinding().to(ModeChoiceCoverageControlerListener.class);
 
-        this.bind(ConvergenceDynamicShutdownImpl.class).in(Singleton.class); //TODO: Delete?
-        this.addControlerListenerBinding().to(ConvergenceDynamicShutdownImpl.class);
+        this.bind(DynamicShutdownControlerListenerImpl.class).in(Singleton.class); //TODO: Delete?
+        this.addControlerListenerBinding().to(DynamicShutdownControlerListenerImpl.class);
 
         this.bind(TerminateDynamically.class).in(Singleton.class); //TODO: Delete?
         this.bind(TerminationCriterion.class).to(TerminateDynamically.class);
